Oh boy... we're in for a loooong ride.
I'm currently testing the import facilities provided by some
external libraries, since for some reason, Godot refuse to
provide the importing facilities within runtime...

Their PCK system is completely broken.
It just does not work when trying to load scenes, and the
fact that you have to know, IN ADVANCE, the name of the files
contained in the PCK file is INSANE, when it comes to
user-generated content.

So, GLB import it is, for the moment... This still won't
do the trick for many others things like, for example,
colliders, but one thing a time... The runtime GLTF importer
from @fire run into a wall when trying to import actual GLTF
with PNG textures, since the PNG texture loader isn't prepared.
So only GLB work correctly.

That said, it REALLY works correctly. I'm able to list
the loaded character animations and play them without
issues. YAY !

So the next step will be Inverse Kinematics !
Fortunately, there's already a Godot project for that.
Then I'll try to incorporate the whole thing inside the
ARVR test project and, MAYBE, we'll get some avatar
loading directly from the producer.

Then there's loading levels created with Mozilla Spoke,
because Mozilla Spoke "just works".

And the last steps are networking + audio conferences
using the WebRTC protocol.

At that point, we should have a decent "application".

Then, we'll attack the real meat of this app, which are :
* Character creation tools, with individual skeletons
  animations, separated from the characters meshes
  (just like MANY animation creation tools like MMD, for
   example) and rigging.
  There's some ideas of sampling some of them from
  Mixamo, or similar services, if the user have accounts
  on such services
* Minecraft-like Voxel based levels, so that you can
  actually play with people using your custom avatars;
* Simple programming tools, by providing simple logic
  bricks, similar to RPG Maker.

Other ideas might come meanwhile, but one thing at
a time...

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>