Turns out that the performances are not THAT bad
You just need to compile Godot in "release_debug", in order to get
a serious speed boost.
That means that the model now takes ~5 seconds to load on a
Ryzen 2700X. Could go faster, but that should do the trick for now.

Let's focus on the real meat. Having a 3D character move with
all the IK and facial animations would be very cool. But,
first things first :

* Get a decent static level generated from a Mozilla Spoke
  configuration
* Provide networking with AudioConferencing

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>