Learning Godot networking
With some videos from the maker of Godot Getaway...

Still, this is the kind of networking that might not
make any sense for the current target.
The fact that any RPC call is executed in every
player system feels bogus to me, and makes replacing
the server by a custom made server system more
difficult, IMHO.

Still, the VOIP client/server example uses a similar
system IIRC, so we'll see if we can combine both.
If I can get ~4 players moving around with custom
models and VOIP on a custom map, using this system,
I'll consider it OK for the 0.0alpha1 .

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>