Still following the Network lesson from Yann Burett
So now I "kind of" understand how the whole thing works.
There are still things like "Node groups" that I don't get,
but these won't be important for the moment.

I guess I should be able to get a multiplayer cube
world today or tomorrow.
Then I'll try to run this on Windows and have another
try at the VOIP examples. With both things working, the
last step would be OpenXR and we should be good to go
for a first alpha try.

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>