VOIP : Not working edition
Packets are sent, packets are received, if I configure the
microphone to output audio on Master, I can hear myself BUUUT
things are not working...

The thing is, I have ZERO information about the audio stack,
when the program is running. I can still ping the Scene state,
but that's it. Added audio channels are not visible, so I have
no idea about what's happening... Just that it doesn't work,
while the VOIP example IS working.

Hmm... I might still consider this as "almost working" and
get to the next steps, and go back to it before releasing
the whole thing, instead of fighting endlessly against something
that is known to be working but isn't for "reasons".

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>