Godot animator is just wonderful
The animation took just a few lines of code, in order
to convert the keyboard input to a joystick like input.

And that's it. The animation blender in Godot
"Just Works" for tihs project. And that's all I need
for now.

There's no crouch, nor jump handling for now.

The next step will be OpenXR. There's already a flag
inside the AnimationTree that freeze the head, arms
and the hips in order to let the Inverse Kinematics
handle them, when coupled with XR inputs.

So the last setps are now :
* Add the OpenXR plugin on Windows
* Implement Inverse Kinematics to move the arms in a
  "plausible" way, so that the character doesn't get
  deformed when moving the arms and the head.
* Compile an independant Windows and Linux binary
* Change the model of the character, and its
  AnimationPlayer with another model that has the same
  animations names, instantiated using a GLB downloaded
  on the network.

That said, if the last step doesn't really work
that well, it will still be OK. I'll release the
0.0alpha1 in that shape.

%aybe just in time for the holidays ?

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>