OpenXR: First support
Turns out that mapping the whole headset, displaying
the the scene inside it was a breeze. However, the
joysticks are not working.

Yet, the A button of left controller was mappable
from Godot action editor...

The next step will be hacking the library quickly
to get basic support for the two joysticks, and then
see about adding a nice action editor to Godot for
OpenXR actions.

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>