OpenXR working edition : With thumbsticks support on Valve Knuckles
The... ARVRCamera is setup in reverse on the main
character, so the controls seem reversed... But
they're working as intended. It's just I'm too
stoopid to get a character controller setup
correctly with VR for the moment.

Still, the VR headset display works, the rotation of the head work,
the thumbsticks work and the buttons work.
I still haven't tested hands tracking with THIS project, but since
it works wiht the "Godot VR Setup" tutorial project, I don't know
why it would not here.

So... Next step : Inverse Kinematics !

... And a decently setup character controller when using VR.

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>