IK: Broken edition
The fact that Mixamo apply some retarded scaling,
that fuck up a lot of things in the process, seems
to pose some serious problems here and there.

That combined with the fact that I have no idea how
to use RenIK properly, make for a seriously
broken result...

So, I'll now try to setup IK with a normally sized
skeleton and mesh, using some CC0-like models
provided by Dwango. Then I'll see about adapting
this to the broken models generated by Mixamo, with
the AnimationPlayer enabled.

And by the way : Fuck FBX. Retarded scaling issues
would not happen if Mixamo were using appropriate
formats in 2020.
This breaks in Blender, breaks in Godot, breaks in
a lot of softwares actually.

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>