VR: Now in option and correctly aligned
The VR movement are now "done". The synchronization is still
done after one second, and only done once, but this should
"do the trick" for now.
So the VR character now looks forward after synchronization,
meaning that the whole player (its capsules, colliders, models,
...) absorb the initial rotation and the camera is looking
forward (and not BACKWARDS !).

It's still bad in many ways, since I only synchronize the
Player rotation in multiplayer, and that rotation is only
sampled on the base node. Meaning that if the HMD is rotated,
due to the real player actually turning in its room, I won't
be able to see it. Maybe I could just sample it on the camera
itself and call it a day ?

One seemingly good idea is to attach everything physically
related to the ARVRCamera so that everything it actually
aligned based on the camera... However, the PlayerRigidBody
needs its collider as a direct child, so that won't be possible.
I haven't tried putting the PlayerRigidBody below the
camera and handling all the movement with it... Note to self...
The main problem with this approach is to make sure that the
actual parent (which might be the ARVROrigin here) is moved,
instead of the PlayerRigidBody, else the RigidBody might just
offset itself from the camera, which won't move, and generate
an insane amount of issues...
Moving the origin manually means that the Physics might not
be taken into account correctly... Ugh...

I'd love to be able to redirect the HMD Transform on any
arbitrary node Transform, instead of playing mind games, though.
If the PlayerRigidBody were to react to the actual HMD offset,
the whole thing would just correct itself.

Anyway, it's not called an adventure for nothing.

Oh, and also a new checkbox haven been added to the main menu,
in order to enable "VR Mode" inside the game.

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>