VR Automatic setup: Half working edition
VR... is really starting to get on my nerves...

That said, putting the model behind the ARVRCamera was
a serious mistake, but it's the kind of mistake you
cannot really see, unless you setup a mirror (which
is still isn't done correclty), or if you think about
it.

The main problem with putting the model behind the
camera is that you get a turret character, the kind of
behaviour you can replicate on some MMO game by
rotating the camera around yourself.
So basically your character will never turn its head...
Which isd bad.

Since I was recently able to focus the VR character
forward and have it synchronized with the PlayerRigidBody,
the next step (for alpha 2 perhapas) will be :
* Reset the ARVRCamera to a "good" state (we will first
  take into account the 'good' scenario where the user
  has its headset on)
* Save the current position on every frame and compare
  it with the previous one and the replicate any move
  by :
  1. Moving the Player and then either :
  2.a Resetting the ARVRCamera position
  2.b Apply an offset to the ARVROrigin that equates
      to the current distance between the Player and
      the ARVRCamera

This free moving camera on which you have no control
is still posing issues, when it comes to setup the
camera based on the model... But again, I'm currently
trying to cheat by placing the model behind the camera,
so there's that too...

Meanwhile, I've been able to do an autosetup for the
moment. It kind of work, but some fiddling is still
required to get it right with the current hacks...

This will have to do the trick, as always, since this
is, like, one of the final piece before releasing alpha1.

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>