Turns out that the camera were facing in the right direction
It's just that the arrow from the move gizmo that seems to
point towards the perspective, actually points BACKWARD !

It seems the Godot editor UI is almost completely broken...
however that means that the VRM is facing the right way,
and so are the camera !

Anyway, the auto IK setup json has been changed... but I
still have to commit this one.

Now the last step will be :
* Test this for non VR users
* Download the actual data from the network
* Apply this automatic setup for any remote user, VR or not.

Then we should have the first version of "avatar" download,
even if the setup will be still completely broken for VR
avatars (still they'll look like attached to the head when
tilting).

In which case we should have the first alpha up and running.

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>