Retesting RenIK
RenIK actually work... If you setup at least 4 look at nodes,
one for the head, one for the hips, and one for each hand.

Now, the "hips" have be setup manually. They don't really
correspond to the real hips. If you don't, the character
will look squished.

Rotation of the target influence the rotation of the bones,
so it should be possible to map the ARVRCamera to the head
target, the ARVRControllers to the hands and the main
rotation to the hip, for the moment.

I cannot really unset the hip at the moment, however in
my current setup, I'd love to have the head and the hands
actually rotate the hips automatically, instead of dealing
with this manually... But still, if I can get this kind
of working for alpha1, this would be a huge step forward.

Turns out I figured this out through VSekai conversations
about full-body tracking, which went me going like "...
Wait !? Maybe their IK system is made to deal with all nodes
setup at once !?".
Because, yes, RenIK is an implemtantion of HumanIK for
Godot, by the VSekai team.

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>