Getting nowhere: VR implementation fucking sucks
I'm still VERY much against this whole "sentient" camera.
This is FUCKING RETARDED, in so many ways !

Think about it :
You make a FPS controller. But instead of moving the player,
and just attaching a camera on its head, you move the camera
instead, and try to reattach everything to the camera on
each frame.
Does it sound sane to you ?
Well, that's how VR system actually work on Godot... And
other engines too, it seems.

It's SOOO stupid. Seriously, camera are supposed to be FIXED
to some place. I'm okay with Tilting, a little less about
pivot. But moving around should NOT move the camera. It
should move the character, on which the camera is attached !

And no, offseting the ARVROrigin every time IS NOT a way to
go.

This whole system is stupid and really need an internal
rework... That said, I'd like to at least release SOMETHING...

The last straw being that the camera position is ALWAYS
relative to last "centering" call. So you cannot try to
regain distance between the player and the camera, if the
camera is moving away, because the camera position is
ALWAYS relative to your current room and the latest
centering call, regardless of the player position or
whatever.
I might still have to try with absolute positions
and see how it kind of work, but I don't think it's going
to work well.

Also, RenIK failed on two other models, so it's unusable.
The only model on which it kind of worked was the KemoonoFriends-Tora
model, but it still could not pull the arms correctly and
had this noodle arm issue, where the forearms try to fold
on themselves when the target was too near, instead of
flexing the elbows back to touch the targeet with the hands.

So, yeah, back to ZombieIK with a camera system that suck
ass. Yay...
No ETA at this point, I need to focus on anothe project
before resuming this one, and I have no idea on how I'm
going to settle these issues.

Meanwhile, WebXR has been added to Godot, so I might
add WebXR support ASAP, once OpenXR support is handled
remotely correctly.
I'm not holding my breath for WebXR, since Linux implementations
are unavailable.

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>