Added a rotating 3D View for models preview
Again, a lot of preparations is still needed to put the
camera at the right position, so that the viewed model
automatically take all the height.

Anyway, this should prove useful to have a deeper look
at the models before instantiating them.

I'm sitll worried by the fact that there's no easy way to
ping the actual amount of video memory used by the program,
which might prove to be problematic when trying to throw
a preview of a model of that small cute synth model that
takes 3GB of VRAM because
"4K Textures ! ON EVERY SINGLE PIXEL !"

By the way, I got my Sketchfab App ID... Incoming :
* Sketchfab download support !
* Advanced Spoke map support !

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>