GLB to GLTF converter : First conversion working
So, in the end, converting GLTF files to GLB is not that
difficult, so I wrote a GDScript that actually do the
conversion.
The next step will be Zip files handling through Godot...
Which should not be a problem since Godot already knows
how to handle Zip files, but it seems that the API to handle
such files is only accessible through a few entrypoints,
like load_resource_pack.

load_resource_pack is too dangerous, since it can easily
overwrite key parts of the projects without warning. It
would be better if, at least, a relative dirpath could be
provided to the function.

Anyway, the conversion works for the moment, but I still
need to be able to convert Zipped GLTF files on the fly.

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>