GLTF to GLB: Handling buffers embedded inside GLTF files files !
Because Blender decided to have an export function that just
does that, embed everything into a single buffer, just like a
GLB, but the buffer is encoded in base64 inside the buffer URI.

I also tested the ability to access user:// GLTF files and convert
them back to GLB, and it works.

So : Yay ! It works !

The last step will be to test this against Sketchfab files.

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>