Sketchfab viewer: Now with actual models !
YES !

IT WORKS !

It's still extra clunky, with zero progression bar or
anything telling you that it's downloading, converting or
failing to do anything...
But still, it works ! I got the models loaded correctly !

YES !

So, the next step will be some cleaning, in order to have
cleaner interfaces, that provide actual progression info,
and showing the actual progress to the user.

And then, invoking the model inside the scene, after making
sure that everything is scaled back, in order to avoid summoning
monsters inside the world.

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>