Don't drink too much coffee edition
I'm trying to finish the UI of the Sketchfab viewer, and then
work on the inventory system, that should list all the models
you can import inside the world.

Then... I don't know... Should I do it GMod style, with a
specific "spawn" gun ? I guess I'll go this route for now,
as this might the simplest option.

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>