Sketchfab viewer : Adding actual filters
It's still kind of clunky, but it works !

I also fixed an issue with the GLTF to GLB converter.
Turns out that you can define a bufferView without
byteOffset... I'm surprised by the number of
tools... ? that are trying to cut corners.
I'm starting to wonder if people are not handcrafting
their GLTF.... I hope not...

Still, I guess I should just implement actual test cases,
and implement the converter so that it follows the norm,
and every potential corner case.

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>